ka.functional: Adds retraction of composition.
A composition can be interpreted as a sequence of two steps (say, `putSocks`
then `putShoes`). Retracting (i.e. undoing) this sequence means retracting the
second step (`removeShoes`), then the first (`removeSocks`).

This means of course that a composition can only be retracted if its steps are.
Translated in terms of concepts, we have (where `=>` is to be read 'implies'):

- RetractableFunction(F) && RetractableFunction(G) => RetractableFunction(composition_t<G, F>)

- IsomorphicAction(F) && IsomorphicAction(G) => IsomorphicAction(composition_accu_t<G, F>)

(a retractable action is always isomorphic due to the domain and codomain being the same)

Change-Id: Id4ad76160c0475685f38c124766f673544564fb7
Reviewed-on: http://gerrit.aldebaran.lan/92411
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: matthieu.paindavoine <matthieu.paindavoine@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: jmonnon <jmonnon@aldebaran.com>