qi.MessageDispatcher: Reduces memory usage by removing recipient ids that are not needed anymore.
In MessageDispatcher, when all handlers for a recipient were
removed, the recipient ID would still be left in the recipients map,
associated with an empty handlers map. As objects are
serialized/deserialized, they add new recipient IDs to the map but
never remove them. So the map grows and never shrinks even long after
the objects have been destroyed.

This patch removes the recipient ID from the map once there is no more
handler for it.

Fixes #45064.

Change-Id: I8548b5e057b5c3681626aae172c7eecb655dd666
Reviewed-on: http://gerrit.aldebaran.lan/96988
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>