qi.ServiceDirectoryProxy: Tunes down a log warning.
Fixes #45095.

The service directory proxy code logs a warning when it fails to attach
to the service directory, explaining that it's going to try attaching
again after a delay. This warning often happens at the start of NAOqi
when the public gateway starts faster than the public service directory
and therefore fails to attach to it. This tends to alert the testers for
no good reason.

This patch just tunes down the log from warning to verbose.

Change-Id: I2dae4865c35b662155635019dc697c6991fd0f47
Reviewed-on: http://gerrit.aldebaran.lan/97024
Reviewed-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jlamotte <jlamotte@aldebaran.com>
Tested-by: vincent.palancher <vincent.palancher@softbankrobotics.com>
Reviewed-by: jmonnon <jmonnon@aldebaran.com>